added comments

+++ src/posts/routers/handlers/delete-post.handler.ts
@@ -4,12 +4,17 @@ import {postsService} from "../../aplication/posts.service";
 
 export async function deletePostHandler(req: Request<{id: string}>, res: Response) {
     try {
+        const post = await postsService.findById(req.params.id);
+        if(!post) {
+            res.sendStatus(HttpStatus.NotFound)
+            return;
+        }
         const isSuccessful = await postsService.delete(req.params.id);
         if (isSuccessful) {
             res.sendStatus(HttpStatus.NoContent);
             return;
         }
-        res.sendStatus(HttpStatus.NotFound)
+        res.sendStatus(HttpStatus.InternalServerError)
     } catch {
         res.sendStatus(HttpStatus.InternalServerError)
     }
